qt4, core: single MarkupParser
