+++ emesene/gui/gtkui/__init__.py
@@ -58,7 +58,6 @@ def gtk_main(Controller):
     import Login
     import MainMenu
     import MainWindow
-    import MarkupParser
 
     try:
         import InfoBar
